add all except profileFragment and chatFragment

+++ app/src/main/java/com/douglas/jointlyapp/ui/base/BaseProfileView.java
@@ -1,5 +1,5 @@
 package com.douglas.jointlyapp.ui.base;
 
 public interface BaseProfileView<T> {
-    void onSuccess(T user, int countUserFollowers, int initiativeCreated, int initiativeJoined);
+    void onSuccess(T user, long countUserFollowers, int initiativeCreated, int initiativeJoined);
 }
